+++ notify_send.py
@@ -50,7 +50,7 @@ SCRIPT_NAME = 'notify_send'
 SCRIPT_AUTHOR = 's3rvac'
 
 # Version of the script.
-SCRIPT_VERSION = '0.8 (dev)'
+SCRIPT_VERSION = '0.8'
 
 # License under which the script is distributed.
 SCRIPT_LICENSE = 'MIT'